refactor: update logLevelType filename
