+++ apps/desktop/src/services/electron-log.service.ts
@@ -2,7 +2,7 @@ import * as path from "path";
 
 import log from "electron-log";
 
-import { LogLevelType } from "@bitwarden/common/enums/logLevelType";
+import { LogLevelType } from "@bitwarden/common/enums/log-level-type";
 import { ConsoleLogService as BaseLogService } from "@bitwarden/common/services/consoleLog.service";
 
 import { isDev } from "../utils";
