+++ libs/common/src/abstractions/log.service.ts
@@ -1,4 +1,4 @@
-import { LogLevelType } from "../enums/logLevelType";
+import { LogLevelType } from "../enums/log-level-type";
 
 export abstract class LogService {
   debug: (message: string) => void;
